+++ config/constants.php
@@ -1,4 +1,8 @@
 <?php
 
 // The site-wide hash key; do not change this because it's used for passwords!
-define('HASH_KEY', 'catsFLYhigh2000miles');
+//this is for more general encryption
+define('HASH_GENERAL_KEY', 'MixItUp2000');
+
+//this is for database passwords only
+define('HASH_PASSWORD_KEY', 'catsFLYhigh2000miles');