+++ controllers/user.php
@@ -22,8 +22,8 @@ class User extends Controller{
      {
         $data = array();
         $data['login'] = $_POST['login'];
-        $data['password'] = md5($_POST['password']);
-        $data['role'] = $_POST['role'];
+        $data['password'] = $_POST['password'];
+        $data['role'] = $_POST['role']; 
 
         // @TODO do your error checking
 
@@ -41,8 +41,7 @@ class User extends Controller{
         $data = array();
         $data['id'] = $id;
         $data['login'] = $_POST['login'];
-        $data['password'] = md5($_POST['password']);
-        // $data['password'] = $_POST['password'];
+        $data['password'] = $_POST['password'];
         $data['role'] = $_POST['role'];
 
         // @TODO do your error checking