+++ models/login_model.php
@@ -9,11 +9,12 @@ class Login_Model extends Model
     
     public function run()
     {
+        
         $sth = $this->db->prepare("SELECT id, role FROM users WHERE 
-        login = :login AND password = MD5(:password)");
+        login = :login AND password = :password");
         $sth->execute(array(
             ':login' => $_POST['login'],
-            ':password' => $_POST['password']
+            ':password' => Hash::create('md5', $_POST['password'] , HASH_PASSWORD_KEY)
         ));
 
         $data = $sth->fetch();