+++ config/paths.php
@@ -1,3 +1,8 @@
 <?php
 
-define('URL', 'http://localhost/mvc/');// defines 'URL' as a constant and defines it as 'http://localhost/php-mvc-tutorial/' This is used in the header.php file ensure that the absolute path is always used  for the respective files
+//Always provide a trailing slash (/) after a path
+
+define('URL', 'http://localhost/mvc/');
+
+
+define('LIBS', 'libs/');