+++ libs/class.model.php
@@ -2,7 +2,7 @@
  class Model {
      function __construct()
      {
-         $this->db = new Database(); //models often use databases. Created one during MVC Tut. p3
+         $this->db = new Database(DB_TYPE, DB_HOST, DB_NAME, DB_USER, DB_PASS); //models often use databases. Created one during MVC Tut. p3
      }
  }