+++ models/user_model.php
@@ -21,31 +21,56 @@ class User_Model extends Model
     }
     public function create($data)//CREATE
     {
-        $sth = $this->db->prepare( 'INSERT INTO users 
-        (`login`,`password`,`role`)
-        VALUES(:login,:password,:role) ');
-        $sth->execute(array(
-            ':login' => $data['login'],
-            ':password' => Hash::create('md5', $data['password'], HASH_PASSWORD_KEY),
-            ':role' => $data[ 'role']
+        $this->db->insert('users', array(
+            'login' => $data['login'],
+            'password' => Hash::create('md5', $data['password'], HASH_PASSWORD_KEY),
+            'role' => $data['role']
         ));
+
+        // $sth = $this->db->prepare( 'INSERT INTO users 
+        // (`login`,`password`,`role`)
+        // VALUES(:login,:password,:role) ');
+
+        // $sth->execute(array(
+        //     ':login' => $data['login'],
+        //     ':password' => Hash::create('md5', $data['password'], HASH_PASSWORD_KEY),
+        //     ':role' => $data[ 'role']
+        // ));
     }
 
 
 
-    public function editSave($data,$id) //UPDATE
+    public function editSave($data) //UPDATE
     {
-        $sth = $this->db->prepare( 'UPDATE users 
-        SET
-         `login` = :login,
-         `password` = :password,
-         `role` = :role
-         WHERE id = :id');
-         $array=array(':id'=>$id,
-                      ':login'=>$data['login'],
-                       ':password'=>Hash::create('md5', $data['password'], HASH_PASSWORD_KEY),
-                      ':role' => $data['role']);
-        $sth->execute($array);
+        $postData = array(
+                ':login'=>$data['login'],
+                ':password'=>Hash::create('md5', $data['password'], HASH_PASSWORD_KEY),
+                ':role' => $data['role']
+            );
+
+       $sth = $this->db->update('users', $postData, "`id` = {$data['id']}");
+
+
+ 
+
+
+        $sth->execute();
+
+        //my code
+
+        // $sth = $this->db->prepare( 'UPDATE users 
+        // SET
+        //  `login` = :login,
+        //  `password` = :password,
+        //  `role` = :role
+        //  WHERE id = :id');
+        //  $array=array(':id'=>[$id],
+        //               ':login'=>$data['login'],
+        //                ':password'=>Hash::create('md5', $data['password'], HASH_PASSWORD_KEY),
+        //               ':role' => $data['role']);
+        // $sth->execute($array);
+
+        //old code
 
         // $sth->execute(array(
         //     ':id' => ['id'],