+++ CHANGELOG.md
@@ -269,7 +269,17 @@ were.
   
                     m   libs/class.models.php - Modified
                             ->changed the Database class creation to pass the db constants from config/database.php to the construct of Database at /libs/class.database.php as variables of the same name
+                 - GIT - Committed.  @1623  
+        @1624
+        Part 6: New insert and update methods 
+            (to make them more universal and useful)
+        https://youtu.be/JmPgJXS7uxA?list=PL7A20112CF84B2229&t=1421
 
-        
-                            
+                    m   /libs/class.database.php
+                            ->Added two new public functions: 
+                            insert($table,$data) 
+                                        &
+                             update($table, $data, $where)      
+                             ->WHAT DO THEY DO
 
+                     m   /libs/class.database.php